ScheduleBlock: put in separate file
