+++ CMakeLists_files.cmake
@@ -198,6 +198,7 @@ if(ENABLE_ECL_INPUT)
     src/opm/input/eclipse/Schedule/RPTConfig.cpp
     src/opm/input/eclipse/Schedule/RSTConfig.cpp
     src/opm/input/eclipse/Schedule/Schedule.cpp
+    src/opm/input/eclipse/Schedule/ScheduleBlock.cpp
     src/opm/input/eclipse/Schedule/ScheduleDeck.cpp
     src/opm/input/eclipse/Schedule/ScheduleGrid.cpp
     src/opm/input/eclipse/Schedule/ScheduleRestartInfo.cpp
@@ -1273,6 +1274,7 @@ if(ENABLE_ECL_INPUT)
        opm/input/eclipse/Schedule/RPTConfig.hpp
        opm/input/eclipse/Schedule/RSTConfig.hpp
        opm/input/eclipse/Schedule/Schedule.hpp
+       opm/input/eclipse/Schedule/ScheduleBlock.hpp
        opm/input/eclipse/Schedule/ScheduleDeck.hpp
        opm/input/eclipse/Schedule/ScheduleGrid.hpp
        opm/input/eclipse/Schedule/ScheduleRestartInfo.hpp
