+++ src/opm/input/eclipse/Schedule/ScheduleDeck.cpp
@@ -40,118 +40,6 @@
 
 namespace Opm {
 
-
-ScheduleBlock::ScheduleBlock(const KeywordLocation& location, ScheduleTimeType time_type, const time_point& start_time) :
-    m_time_type(time_type),
-    m_start_time(start_time),
-    m_location(location)
-{
-}
-
-std::size_t ScheduleBlock::size() const {
-    return this->m_keywords.size();
-}
-
-void ScheduleBlock::push_back(const DeckKeyword& keyword) {
-    this->m_keywords.push_back(keyword);
-}
-
-std::vector<DeckKeyword>::const_iterator ScheduleBlock::begin() const {
-    return this->m_keywords.begin();
-}
-
-std::vector<DeckKeyword>::const_iterator ScheduleBlock::end() const {
-    return this->m_keywords.end();
-}
-
-const DeckKeyword& ScheduleBlock::operator[](const std::size_t index) const {
-    return this->m_keywords.at(index);
-}
-
-const time_point& ScheduleBlock::start_time() const {
-    return this->m_start_time;
-}
-
-const std::optional<time_point>& ScheduleBlock::end_time() const {
-    return this->m_end_time;
-}
-
-ScheduleTimeType ScheduleBlock::time_type() const {
-    return this->m_time_type;
-}
-
-void ScheduleBlock::end_time(const time_point& t) {
-    this->m_end_time = t;
-}
-
-bool ScheduleBlock::operator==(const ScheduleBlock& other) const {
-    return this->m_start_time == other.m_start_time &&
-           this->m_end_time == other.m_end_time &&
-           this->m_location == other.m_location &&
-           this->m_time_type == other.m_time_type &&
-           this->m_keywords == other.m_keywords;
-}
-
-void ScheduleBlock::dump_time(time_point current_time, DeckOutput& output) const {
-    if (this->m_time_type == ScheduleTimeType::START)
-        return;
-
-    if (this->m_time_type == ScheduleTimeType::DATES) {
-        TimeStampUTC ts(TimeService::to_time_t(this->start_time()));
-        auto ecl_month = TimeService::eclipseMonthNames().at(ts.month());
-        std::string dates_string = fmt::format(R"(
-DATES
-   {} '{}' {} /
-/
-)", ts.day(), ecl_month, ts.year());
-        output.write_string(dates_string);
-    } else {
-        auto seconds = std::chrono::duration_cast<std::chrono::seconds>(this->start_time() - current_time);
-        double days = seconds.count() / 86400.0;
-        std::string tstep_string = fmt::format(R"(
-TSTEP
-   {} /
-)", days);
-        output.write_string(tstep_string);
-    }
-}
-
-
-void ScheduleBlock::dump_deck(DeckOutput& output, time_point& current_time) const {
-    this->dump_time(current_time, output);
-    if (!this->end_time().has_value())
-        return;
-
-    for (const auto& keyword : this->m_keywords)
-        keyword.write(output);
-
-    current_time = this->end_time().value();
-}
-
-
-const KeywordLocation& ScheduleBlock::location() const {
-    return this->m_location;
-}
-
-
-ScheduleBlock ScheduleBlock::serializationTestObject() {
-    ScheduleBlock block;
-    block.m_time_type = ScheduleTimeType::TSTEP;
-    block.m_start_time = TimeService::from_time_t( asTimeT( TimeStampUTC( 2003, 10, 10 )));
-    block.m_end_time = TimeService::from_time_t( asTimeT( TimeStampUTC( 1993, 07, 06 )));
-    block.m_location = KeywordLocation::serializationTestObject();
-    block.m_keywords = {DeckKeyword::serializationTestObject()};
-    return block;
-}
-
-std::optional<DeckKeyword> ScheduleBlock::get(const std::string& kw) const {
-    for (const auto& keyword : this->m_keywords) {
-        if (keyword.name() == kw)
-            return keyword;
-    }
-    return {};
-}
-
 /*****************************************************************************/
 
 struct ScheduleDeckContext {
